+++ src/Search/SearchBar.tsx
@@ -7,7 +7,7 @@ import {
 } from "office-ui-fabric-react";
 import { RouteComponentProps } from "react-router";
 import { gotoSearch, SearchRouteProps, getSearchMatch } from "./Routing";
-import { Subject } from "../subject/model/Subject";
+import { Subject } from "../model/Subject";
 
 const styles = mergeStyleSets({
   search: {
